Improve enum naming
The protobuf guide recommends that enums be upper snakecase.

The buf-linter enforces this and additionally requires that
all enum values be prefixed with the enum name, e.g.:

```
enum ErrorCode {
  ERROR_CODE_UNSPECIFIED = 0;
  ERROR_CODE_NOT_FOUND = 1;
  ERROR_CODE_ACCESS_DENIED = 2;
  ERROR_CODE_NOT_IMPLEMENTED = 3;
}
```

The reasoning is explained here:
https://docs.buf.build/lint-rules#enum_value_prefix

In the past protobuf.cr would unfortunately translate
this style to rather unpleasant crystal identifiers, e.g.:

```
enum ErrorCode
  ERRORCODEUNSPECIFIED = 0
  ERRORCODENOTFOUND = 1
  ERRORCODEACCESSDENIED = 2
  ERRORCODENOTIMPLEMENTED = 3
end
```

This patch enforces upper snakecase for enums and
strips the prefix when it matches the enum-name.

For the above example it produces:

```
enum ErrorCode
  UNSPECIFIED = 0
  NOT_FOUND = 1
  ACCESS_DENIED = 2
  NOT_IMPLEMENTED = 3
end
```